requirements.txt
praat-parselmouth
pandas
numpy
plotly
